% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{compute_GHG_cpp}
\alias{compute_GHG_cpp}
\title{Compute GHG Matrix for Score Variance}
\usage{
compute_GHG_cpp(G, Hhalf)
}
\arguments{
\item{G}{Numeric matrix (n x p): Genotype or weighted genotype matrix}

\item{Hhalf}{Numeric matrix (n x k): H^(1/2) projection matrix}
}
\value{
List with components:
\item{GHG}{Matrix (p x p): Score variance matrix}
\item{GHhalf}{Matrix (p x k): Intermediate product G' H^(1/2)}
}
\description{
Efficiently computes the score variance matrix GHG = G'G - G'H H'G
using Armadillo linear algebra library.
}
\details{
This function computes:
GtG = G' * G
GHhalf = G' * H^(1/2)
GHG = GtG - GHhalf * GHhalf'

The C++ implementation provides 2-5x speedup over pure R implementation.
}
\keyword{internal}
