% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_PI_train.R
\docType{package}
\name{get_PI_train}
\alias{get_PI_train}
\title{PI Model Training Functions}
\description{
Functions for training PI (variant-importance score) ensemble models
from pre-annotated case and control variant data.
}
\author{
\strong{Maintainer}: GLOWpipeline Project \email{glow@example.com}

}
