% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variant_filter.R
\name{print.glow_annotation_predicate}
\alias{print.glow_annotation_predicate}
\title{Print Method for Annotation Predicates}
\usage{
\method{print}{glow_annotation_predicate}(x, ...)
}
\arguments{
\item{x}{A \code{glow_annotation_predicate} object.}

\item{...}{Additional arguments (ignored).}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print Method for Annotation Predicates
}
