% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_B_diagnostics.R
\name{print.glow_B_diagnostics}
\alias{print.glow_B_diagnostics}
\title{Print Method for B Diagnostics}
\usage{
\method{print}{glow_B_diagnostics}(x, ...)
}
\arguments{
\item{x}{Object of class "glow_B_diagnostics"}

\item{...}{Additional arguments (unused)}
}
\value{
The object x (invisibly)
}
\description{
S3 print method for glow_B_diagnostics objects.
}
