% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glow_input.R
\name{print.glow_input}
\alias{print.glow_input}
\title{Print Method for glow_input Objects}
\usage{
\method{print}{glow_input}(x, ...)
}
\arguments{
\item{x}{A \code{glow_input} object.}

\item{...}{Additional arguments (ignored).}
}
\value{
Invisible \code{x}.
}
\description{
Print Method for glow_input Objects
}
