% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_B_outlier_detection.R
\name{print.glow_outliers}
\alias{print.glow_outliers}
\title{Print Method for glow_outliers Objects}
\usage{
\method{print}{glow_outliers}(x, max_show = 10, ...)
}
\arguments{
\item{x}{An object of class "glow_outliers"}

\item{max_show}{Maximum number of outliers to display in detail. Default is 10.}

\item{...}{Additional arguments (currently unused)}
}
\value{
Invisibly returns the input object
}
\description{
Prints a summary of detected outliers in a human-readable format.
}
