% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_PI_control_data_loader.R
\name{print.glow_pi_control_data}
\alias{print.glow_pi_control_data}
\title{Print Method for glow_pi_control_data}
\usage{
\method{print}{glow_pi_control_data}(x, ...)
}
\arguments{
\item{x}{Object of class \code{glow_pi_control_data}}

\item{...}{Additional arguments (unused)}
}
\description{
Prints a summary of prepared PI control variant data.
}
