% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_B_data_loader.R
\name{print.glow_training_data}
\alias{print.glow_training_data}
\title{Print Method for glow_training_data}
\usage{
\method{print}{glow_training_data}(x, ...)
}
\arguments{
\item{x}{Object of class \code{glow_training_data}}

\item{...}{Additional arguments (unused)}
}
\description{
Prints a summary of prepared GWAS training data.
}
