% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variant_filter.R
\name{print.glow_variant_filter}
\alias{print.glow_variant_filter}
\title{Print Method for Variant Filter Specifications}
\usage{
\method{print}{glow_variant_filter}(x, ...)
}
\arguments{
\item{x}{A \code{glow_variant_filter} object.}

\item{...}{Additional arguments (ignored).}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print Method for Variant Filter Specifications
}
