% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_PI_case_data_loader.R
\name{summary.glow_pi_case_data}
\alias{summary.glow_pi_case_data}
\title{Summary Method for glow_pi_case_data}
\usage{
\method{summary}{glow_pi_case_data}(object, ...)
}
\arguments{
\item{object}{Object of class \code{glow_pi_case_data}}

\item{...}{Additional arguments (unused)}
}
\description{
Provides a detailed summary of prepared PI case variant data.
}
