% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_B_data_loader.R
\name{summary.glow_training_data}
\alias{summary.glow_training_data}
\title{Summary Method for glow_training_data}
\usage{
\method{summary}{glow_training_data}(object, ...)
}
\arguments{
\item{object}{Object of class \code{glow_training_data}}

\item{...}{Additional arguments (unused)}
}
\description{
Provides a detailed summary of prepared GWAS training data.
}
